Define rake spec task
I ran into an issue with the skeleton when I tried to run my test suite using either `rake spec` or the default rake task.  When I ran the tasks, nothing happened, and it appeared that things failed silently; I was not told that rake didn't know how to build the task 'spec'.

If I ran the test suite with the `rspec` command, the tests ran.

I was able to run the test suite through the rake tasks after defining the spec task in this commit.